+++ benchmarks/express.js
@@ -1,6 +1,6 @@
 'use strict'
 
-const { title, now, print, operations } = require('./utils')
+const { title, now, print, operations } = require('../utils')
 const router = require('express/lib/router')()
 
 title('express benchmark')
